Implement competence test
New user needs to pass it to annotate series and validate annotations.
To pass test one needs to correctly annotate 5 series in a row.

All series are selected to be verified (best_cohens_kappa = 1)
First 3 series are selected to be non-controversial (fleiss_kappa = 1)
Last 2 are selected to be less obvious (fleiss_kappa < 1)